Move linker-related functionality of S2WasmBuilder into its own class
